+++ src/app/cart/services/cart.service.ts
@@ -5,58 +5,66 @@ import { ICartProduct } from '../../shared/types';
   providedIn: 'root'
 })
 export class CartService {
-  products: ICartProduct[] = [];
+  private _cartProducts: ICartProduct[] = [];
   private _isCartListShown: boolean = false;
 
   get isCartListShown(): boolean{
     return this._isCartListShown
   }
 
-  constructor() { }
-
-  public getCartProducts(): ICartProduct[] | []  {
-    return this.products
+  get totalQuantity(): number {
+    return this._cartProducts.reduce((calc: number, cur: ICartProduct) => {
+      calc += cur.quantity
+      return calc
+    }, 0)
   }
 
-
-  public addToCart(productItem: ICartProduct): void {
-    const index = this.findIndex(productItem);
-      index < 0
-        ? this.products.push(productItem)
-        : this.increase(index);
-  }
-
-  public increase(index: number): void {
-    this.products[index].quantity++
-  }
-
-  public decrease(index: number): void {
-    this.products[index].quantity--
-  }
-
-  public findIndex(cartProduct: ICartProduct): number {
-    return this.products.findIndex(product => product.id === cartProduct.id)
-  }
-
-  public deleteCartItem(productItem: ICartProduct) {
-    this.products = this.products.filter(product => product.id !== productItem.id);
-  }
-
-  public getCartSum(): number {
-    return this.products.reduce((calc: number, cur: ICartProduct) => {
+  get totalSum(): number {
+    return this._cartProducts.reduce((calc: number, cur: ICartProduct) => {
       calc += cur.quantity * cur.price
       return calc
     }, 0)
+  }
+
+  get isEmptyCart(): boolean {
+    return !!this.totalSum
+  }
+
+  constructor() { }
+
+  public getCartProducts(): ICartProduct[] | []  {
+    return this._cartProducts
+  }
+
+
+  public addProduct(productItem: ICartProduct): void {
+    const index = this.findIndex(productItem);
+      index < 0
+        ? this._cartProducts.push(productItem)
+        : this.increaseQuantity(index);
+  }
+
+  public increaseQuantity(index: number): void {
+    this._cartProducts[index].quantity++
+  }
+
+  public decrease(index: number): void {
+    this._cartProducts[index].quantity--
+  }
+
+  public findIndex(cartProduct: ICartProduct): number {
+    return this._cartProducts.findIndex(product => product.id === cartProduct.id)
+  }
+
+  public removeProduct(productItem: ICartProduct) {
+    this._cartProducts = this._cartProducts.filter(product => product.id !== productItem.id);
   }
 
   public toggleCartList(): void {
     this._isCartListShown = !this._isCartListShown
   }
 
-  public getNumberPurchases(): number {
-    return this.products.reduce((calc: number, cur: ICartProduct) => {
-      calc += cur.quantity
-      return calc
-    }, 0)
+  public removeAllProducts(): void {
+    this._cartProducts = []
   }
 }