+++ src/app/product/services/products.service.ts
@@ -2,6 +2,7 @@ import { Injectable } from '@angular/core';
 import productList from '../../shared/product-list';
 import { IProduct } from '../../shared/types';
 import { CartService } from '../../cart/services/cart.service';
+import { ConstantsService } from '../../shared/constants.service';
 
 @Injectable({
   providedIn: 'root'
@@ -10,13 +11,16 @@ export class ProductsService {
 
   constructor(
     private cartService: CartService,
-  ) { }
+    private constantService: ConstantsService
+  ) {
+    console.log('URL', this.constantService.url)
+  }
 
-  public getProducts(): IProduct[]  {
+  public getProducts(): IProduct[] {
     return productList
   }
 
   public addToCart(product: IProduct): void {
-    this.cartService.addToCart({ ...product, quantity: 1})
+    this.cartService.addProduct({ ...product, quantity: 1})
   }
 }