+++ src/app/shared/generator.factory.ts
@@ -1,0 +1,7 @@
+import { GeneratorService } from './generator.service';
+
+export function GeneratorFactory(length: number) {
+  return function(generator: GeneratorService): string {
+    return generator.generate(length).toUpperCase()
+  }
+}